Restrict `LinearFunction` objects to `LSpace`.
This is required for the functor instances from `LinearFunction` to `LinearMap`.

Unfortunately, this means `LinearFunction` is not as lightwight anymore as it
used to be. Also, the free vector constructors `V0`...`V4` can't be functors in
the `LinearFunction` category anymore (since `Vn v` is, at the moment, only a
`VectorSpace` if `v` is a scalar).
This is not too big a deal since the internally equivalent `Tensor s (Vn s) v`
embodies the concept better anyway.