Test case that reveals bug in `dualBasis`.
If the first given vector, in a cartesian system, has a zero somewhere but the
first component (e.g. `(1.3, 0)`) and another vector has a larger first
component, the algorithm will start with a first candidate that's actually
orthogonal to the vector we want to match it with.

This obviously breaks the following Gram-Schmidt process.