Remove the unnecessary `LV` argument to the `FreeLinearSpace` macro.
This argument formerly distinguished between different constructors of the LinearMaps
over the V0...V4 spaces, but this is now using the same constructor (in a nested fashion)
of the space itself, with the consequence that all macro invocations passed the same
`LinearMap` wrapper-constructor. This can as well be inlined in the macro.